requirements.txt
torch==1.9
numpy==1.19.5
lmdb
pyarrow==0.14.0
pyyaml
easydict
ConfigArgParse
tensorboard
torchvision
einops
matplotlib
tqdm
torchsnooper
ema-pytorch
accelerate
soundfile
pandas==1.3.4
transforms3d
scipy
scikit-learn
librosa
omegaconf
sox
rich
ffmpeg-normalize
blobfile
